+++ packages/plugin-pha/src/manifestHelpers.ts
@@ -142,7 +142,7 @@ async function renderPageDocument(routeId: string, serverEntry: string): Promise
     },
   };
   const serverModule = await import(serverEntry);
-  const { value } = await serverModule.renderToEntry(serverContext, {
+  const { value } = await serverModule.renderToHTML(serverContext, {
     documentOnly: true,
     serverOnlyBasename: '/',
     renderMode: 'SSG',